AG-18445 Move the carousel into ag-website-shared as a reusable component
