+++ packages/ag-charts-website/src/components/gallery/components/GalleryExampleLink.module.scss
@@ -64,9 +64,7 @@
 
 // A carousel slide: fixed width so a partly-visible neighbour signals more to scroll to.
 .layout-thumbnail {
-    flex: 0 0 auto;
     width: 240px;
-    scroll-snap-align: start;
 
     .image {
         padding: $spacing-size-3;
